Add explicit information about classification sources
In the database, sources cited in classification comments have been identified
explicitely. This info is transferred to the repos with this PR. While in many
(most?) cases this is redundant, I think this will allow a richer description
of the classification in the future, with a clear separation of bibliography from
prose for the classification comments.