Correcting geocoordinates
The language Vano was assigned to the wrong island.  It's a language of Vanikoro I., traditionally in NW area.  Was down to 4 speakers in my last field trip of 2012;  so yes, it's a moribund language for sure. See http://alex.francois.online.fr/AF-maps-Vanikoro.htm 
(I'll make a separate commit to propose renaming it to "Lovono".)