Changing lg name
The name "Vano" is not recognised by modern speakers.  The native name is "Alavana", but being a moribund language, very few people recognise that form either.  The common way to refer to the language is "Lovono", i.e. the same placename as it is named in the main island's sole surviving language Teanu.  I therefore propose to rename it "Lovono". See http://alex.francois.online.fr/AF-maps-Vanikoro.htm 
(I made a separate change about its spatial coordinates;  I'm not well used to GitHub, so I'm not sure this is the best way to proceed.)