Removed links to Louisiana French
The source used for this entry says that Gulf Coast French isn't a variety of Louisiana French. I checked the edit history of the Louisiana French article on Wikipedia and a link to the Gulf Coast French Glottolog entry was added before the links to Wikipedia and Wikidata's Louisiana French articles were added to the Gulf Coast French languoid. It is purely citogenesis.